Move nftsFromOperations to coin-framework
